some fix

+++ app/Http/Controllers/SiteController.php
@@ -9,17 +9,23 @@ use Illuminate\Http\Request;
 use App\Site;
 use App\Http\Requests;
 use Input;
+use DB;
 
 class SiteController extends Controller
 {
     //
+    private function getDefaultTemplate()
+    {
+        $DefaultTemplate = DB::table('templates')->first();
+        return $DefaultTemplate->id ;
+    }
 
     public function activetemp($id){
         $site_id=Auth::user()->site->id;
         $row=Site::find($site_id);
         $row->template_id=$id;
         $row->save();
-        return view('dashboard');
+        return redirect('/');
 
 
     }
@@ -48,7 +54,7 @@ class SiteController extends Controller
      public function store(Request $request)
      {
         $this->validate($request, [
-            'subdomain' => 'required|max:255',
+            'subdomain' => 'required|max:255|unique:sites',
             'color' => 'required',
             'primary_color' => 'required',
             'secondry_color' => 'required',
@@ -58,6 +64,7 @@ class SiteController extends Controller
         $site = new Site($request->all());
         $site->id = Auth::user()->id;
         $site->user_id = Auth::user()->id;
+        $site->template_id =SiteController::getDefaultTemplate();; 
         $imagePath='';
         if(Input::hasFile('background_image')){  
                 $file = Input::file('background_image');
@@ -70,7 +77,7 @@ class SiteController extends Controller
         $site->background_image = $imagePath;
         if($site->addSite($site)) 
         {
-            Session::put('site_id', $site->id);
+            // Session::put('site_id', $site->id);
             // return  redirect('/site');
             return redirect()->action('TemplateController@index');
     
@@ -99,8 +106,10 @@ class SiteController extends Controller
         ]);
         $old_imag_name = $site['attributes']['background_image'];
 
-        $imagePath='';
+        $imagePath=$old_imag_name;
+        $has_new_file = 0;
         if(Input::hasFile('background_image')){  
+                 $has_new_file = 1;
                 $file = Input::file('background_image');
                 $extension = $file->getClientOriginalExtension();
                 $file = $file->move(public_path().'/assets/images/'.$site->subdomain.'/backgrounds/',time().'.'.$extension);
@@ -115,9 +124,17 @@ class SiteController extends Controller
             'secondry_color' => $request->all()['secondry_color'],
             'body_type' => $request->all()['body_type'],
             'background_image' => $imagePath,
+
         ]))
        {
-            unlink(public_path().$old_imag_name);    
+        if($has_new_file == 1)
+        {
+                try {
+                    unlink(public_path().$old_imag_name);    
+                }
+                catch (\Exception $e) {
+                }
+        }
             return redirect('/dashboard');
        }
        return back();
